added ElseDo and ElseDo Async Tests
